refactor(session): decompose _managed_headless_session_lineage.py
Move per-record creation/projection/anchor helpers to _records.py,
runner observation helpers to _runner.py, and index path helpers
to _indexes.py. _managed_headless_session_lineage.py retains
the conflict/CAS mismatch errors and the DefaultManagedHeadlessSessionLineageStore
class plus the unaccounted _validate_anchor_identity/_require_cas/
_read_bounded helpers. Existing canonical-path imports keep working
via re-export facade.

Refs: #4664
