+++ src/autoskillit/execution/session/_managed_headless_session_lineage.py
@@ -549,330 +549,30 @@ class DefaultManagedHeadlessSessionLineageStore:
             return updated
 
 
-def _new_lineage(
-    *,
-    launch_id: str,
-    decision: NativeShellCaptureDecision,
-    backend: str,
-    session_kind: ManagedHeadlessSessionKind,
-    lineage_anchor: Path,
-    anchor_device: int,
-    anchor_inode: int,
-    dispatch_id: str | None,
-) -> ManagedHeadlessSessionLineage:
-    identity = {
-        "schema_version": MANAGED_HEADLESS_SESSION_LINEAGE_SCHEMA_VERSION,
-        "launch_id": launch_id,
-        "decision": decision.to_dict(),
-        "backend": backend,
-        "session_kind": session_kind.value,
-        "lineage_anchor": str(lineage_anchor),
-        "anchor_device": anchor_device,
-        "anchor_inode": anchor_inode,
-    }
-    lineage_digest = _digest(identity)
-    provisional = ManagedHeadlessSessionLineage(
-        launch_id=launch_id,
-        decision=decision,
-        backend=backend,
-        session_kind=session_kind,
-        lineage_anchor=str(lineage_anchor),
-        anchor_device=anchor_device,
-        anchor_inode=anchor_inode,
-        lineage_digest=lineage_digest,
-        generation=0,
-        record_digest="0" * 64,
-        dispatch_id=dispatch_id,
-    )
-    return replace(provisional, record_digest=_digest(_record_payload(provisional)))
-
-
-def _next_generation(
-    lineage: ManagedHeadlessSessionLineage,
-) -> ManagedHeadlessSessionLineage:
-    provisional = replace(
-        lineage,
-        generation=lineage.generation + 1,
-        record_digest="0" * 64,
-    )
-    return replace(provisional, record_digest=_digest(_record_payload(provisional)))
-
-
-def _creation_projection(lineage: ManagedHeadlessSessionLineage) -> tuple[object, ...]:
-    return (
-        lineage.launch_id,
-        lineage.decision,
-        lineage.backend,
-        lineage.session_kind,
-        lineage.lineage_anchor,
-        lineage.anchor_device,
-        lineage.anchor_inode,
-        lineage.dispatch_id,
-    )
-
-
-def _resolve_anchor(lineage_anchor: Path) -> tuple[Path, int, int]:
-    supplied = Path(lineage_anchor).expanduser()
-    if not supplied.is_absolute():
-        raise ValueError("Managed lineage anchor must be absolute")
-    try:
-        anchor = supplied.resolve(strict=True)
-        stat_result = anchor.stat()
-    except OSError as exc:
-        raise ValueError("Managed lineage anchor is unavailable") from exc
-    if not anchor.is_dir():
-        raise ValueError("Managed lineage anchor must be a directory")
-    return anchor, stat_result.st_dev, stat_result.st_ino
-
-
-def _prepare_root(anchor: Path) -> Path:
-    current = anchor
-    for component in _NAMESPACE.parts:
-        current = current / component
-        if current.exists() and current.is_symlink():
-            raise ValueError("Managed lineage namespace cannot contain symlinks")
-        current.mkdir(mode=0o700, exist_ok=True)
-        if current.is_symlink() or not current.is_dir():
-            raise ValueError("Managed lineage namespace is not a regular directory")
-    for relative in (
-        Path(_RECORDS_DIR),
-        Path(_INDEXES_DIR) / _FINAL_NATIVE_INDEX,
-        Path(_INDEXES_DIR) / _DISPATCH_INDEX,
-    ):
-        directory = current / relative
-        directory.mkdir(mode=0o700, parents=True, exist_ok=True)
-        if directory.is_symlink() or not directory.is_dir():
-            raise ValueError("Managed lineage namespace is not a regular directory")
-    return current
-
-
-@contextmanager
-def _store_lock(root: Path, *, exclusive: bool) -> Iterator[None]:
-    lock_path = root / _LOCK_FILENAME
-    flags = os.O_CREAT | os.O_RDWR
-    if hasattr(os, "O_NOFOLLOW"):
-        flags |= os.O_NOFOLLOW
-    fd = os.open(lock_path, flags, 0o600)
-    try:
-        fcntl.flock(fd, fcntl.LOCK_EX if exclusive else fcntl.LOCK_SH)
-        yield
-    finally:
-        fcntl.flock(fd, fcntl.LOCK_UN)
-        os.close(fd)
-
-
-def _record_path(root: Path, launch_id: str) -> Path:
-    # Construction validates the identity before this path can be written.
-    if (
-        not isinstance(launch_id, str)
-        or len(launch_id) != 32
-        or any(character not in "0123456789abcdef" for character in launch_id)
-    ):
-        raise ValueError("Invalid launch_id")
-    return root / _RECORDS_DIR / f"{launch_id}.json"
-
-
-def _write_record(path: Path, lineage: ManagedHeadlessSessionLineage) -> None:
-    atomic_write(
-        path,
-        _canonical_json(_record_to_dict(lineage)),
-        strict_durability=True,
-    )
-
-
-def _read_record(path: Path) -> ManagedHeadlessSessionLineage:
-    raw = _read_bounded(path)
-    value = _strict_json_load(raw)
-    if _canonical_json(value).encode("utf-8") != raw:
-        raise ValueError("Managed lineage record is not canonical JSON")
-    return _lineage_from_dict(value)
-
-
-def _read_runner_markers(
-    root: Path,
-    reference: ManagedHeadlessSessionLineageRef,
-    lineage: ManagedHeadlessSessionLineage,
-) -> tuple[NativeShellCaptureObservation, ...]:
-    directory_flags = os.O_RDONLY | getattr(os, "O_DIRECTORY", 0)
-    nofollow = getattr(os, "O_NOFOLLOW", 0)
-    root_fd = os.open(root, directory_flags | nofollow)
-    observations_fd = -1
-    launch_fd = -1
-    try:
-        try:
-            observations_fd = os.open(
-                _RUNNER_OBSERVATIONS_DIR,
-                directory_flags | nofollow,
-                dir_fd=root_fd,
-            )
-            launch_fd = os.open(
-                reference.launch_id,
-                directory_flags | nofollow,
-                dir_fd=observations_fd,
-            )
-        except FileNotFoundError:
-            return ()
-        parsed: list[NativeShellCaptureObservation] = []
-        for name in sorted(os.listdir(launch_fd))[:_MAX_RUNNER_MARKERS]:
-            if not name.endswith(".json") or "/" in name or name in {".", ".."}:
-                continue
-            marker_fd = -1
-            try:
-                marker_fd = os.open(
-                    name,
-                    os.O_RDONLY | nofollow,
-                    dir_fd=launch_fd,
-                )
-                metadata = os.fstat(marker_fd)
-                if (
-                    not stat.S_ISREG(metadata.st_mode)
-                    or metadata.st_nlink != 1
-                    or metadata.st_size > _MAX_RUNNER_MARKER_BYTES
-                ):
-                    continue
-                raw = os.read(marker_fd, _MAX_RUNNER_MARKER_BYTES + 1)
-            except OSError:
-                continue
-            finally:
-                if marker_fd >= 0:
-                    os.close(marker_fd)
-            try:
-                marker = _strict_json_load(raw)
-                if _canonical_json(marker).encode("utf-8") != raw:
-                    continue
-                if not isinstance(marker, dict) or set(marker) != {
-                    "schema_version",
-                    "launch_id",
-                    "lineage_digest",
-                    "observation",
-                }:
-                    continue
-                if (
-                    marker["schema_version"] != MANAGED_HEADLESS_SESSION_LINEAGE_SCHEMA_VERSION
-                    or marker["launch_id"] != reference.launch_id
-                    or marker["lineage_digest"] != reference.lineage_digest
-                ):
-                    continue
-                observation = NativeShellCaptureObservation.from_dict(marker["observation"])
-                if observation.attempt_id not in lineage.attempt_ids:
-                    continue
-            except (TypeError, ValueError):
-                continue
-            parsed.append(observation)
-        return tuple(dict.fromkeys(parsed))
-    finally:
-        if launch_fd >= 0:
-            os.close(launch_fd)
-        if observations_fd >= 0:
-            os.close(observations_fd)
-        os.close(root_fd)
-
-
-def _settle_runner_observation(
-    root: Path,
-    reference: ManagedHeadlessSessionLineageRef,
-    observation: NativeShellCaptureObservation,
-) -> None:
-    """Durably consume one marker after its lineage mutation has settled."""
-    marker = {
-        "schema_version": MANAGED_HEADLESS_SESSION_LINEAGE_SCHEMA_VERSION,
-        "launch_id": reference.launch_id,
-        "lineage_digest": reference.lineage_digest,
-        "observation": observation.to_dict(),
-    }
-    marker_name = f"{hashlib.sha256(_canonical_json(marker).encode('utf-8')).hexdigest()}.json"
-    directory_flags = os.O_RDONLY | getattr(os, "O_DIRECTORY", 0)
-    nofollow = getattr(os, "O_NOFOLLOW", 0)
-    root_fd = os.open(root, directory_flags | nofollow)
-    observations_fd = -1
-    launch_fd = -1
-    try:
-        try:
-            observations_fd = os.open(
-                _RUNNER_OBSERVATIONS_DIR,
-                directory_flags | nofollow,
-                dir_fd=root_fd,
-            )
-            launch_fd = os.open(
-                reference.launch_id,
-                directory_flags | nofollow,
-                dir_fd=observations_fd,
-            )
-            os.unlink(marker_name, dir_fd=launch_fd)
-            os.fsync(launch_fd)
-        except FileNotFoundError:
-            return
-    finally:
-        if launch_fd >= 0:
-            os.close(launch_fd)
-        if observations_fd >= 0:
-            os.close(observations_fd)
-        os.close(root_fd)
-
-
-def _index_path(root: Path, index_name: str, key: str) -> Path:
-    if not isinstance(key, str) or not key or "\x00" in key:
-        raise ValueError(f"Invalid managed lineage {index_name} key")
-    if len(key.encode("utf-8")) > 512:
-        raise ValueError(f"Managed lineage {index_name} key is oversized")
-    digest = hashlib.sha256(key.encode("utf-8")).hexdigest()
-    return root / _INDEXES_DIR / index_name / f"{digest}.json"
-
-
-def _write_index(root: Path, index_name: str, key: str, launch_id: str) -> None:
-    path = _index_path(root, index_name, key)
-    atomic_write(
-        path,
-        _canonical_json(
-            {
-                "schema_version": MANAGED_HEADLESS_SESSION_LINEAGE_SCHEMA_VERSION,
-                "key": key,
-                "launch_id": launch_id,
-            }
-        ),
-        strict_durability=True,
-    )
-
-
-def _remove_index(root: Path, index_name: str, key: str) -> None:
-    """Durably remove one index entry while its namespace lock is held."""
-    path = _index_path(root, index_name, key)
-    path.unlink()
-    directory_fd = os.open(path.parent, os.O_RDONLY)
-    try:
-        os.fsync(directory_fd)
-    finally:
-        os.close(directory_fd)
-
-
-def _read_index(root: Path, index_name: str, key: str) -> str:
-    path = _index_path(root, index_name, key)
-    value = _strict_json_load(_read_bounded(path))
-    expected_fields = {"schema_version", "key", "launch_id"}
-    if not isinstance(value, dict) or set(value) != expected_fields:
-        raise ValueError("Invalid managed lineage index")
-    if value["schema_version"] != MANAGED_HEADLESS_SESSION_LINEAGE_SCHEMA_VERSION:
-        raise ValueError("Unsupported managed lineage index schema")
-    if value["key"] != key:
-        raise ValueError("Managed lineage index key mismatch")
-    return _strict_str(value["launch_id"], "launch_id")
-
-
-def _assert_index_available(
-    root: Path,
-    index_name: str,
-    key: str,
-    launch_id: str,
-) -> None:
-    path = _index_path(root, index_name, key)
-    if not path.exists():
-        return
-    indexed_launch_id = _read_index(root, index_name, key)
-    if indexed_launch_id != launch_id:
-        raise ManagedHeadlessSessionLineageConflictError(
-            f"Managed lineage {index_name} identity is already owned"
-        )
-
+# Index/runner/records helpers live in sibling modules; re-exported for
+# existing callers using the canonical _managed_headless_session_lineage path.
+from autoskillit.execution.session._managed_headless_session_lineage_records import (  # noqa: F401
+    _creation_projection,
+    _new_lineage,
+    _next_generation,
+    _prepare_root,
+    _read_record,
+    _record_path,
+    _resolve_anchor,
+    _store_lock,
+    _write_record,
+)
+from autoskillit.execution.session._managed_headless_session_lineage_runner import (  # noqa: F401
+    _read_runner_markers,
+    _settle_runner_observation,
+)
+from autoskillit.execution.session._managed_headless_session_lineage_indexes import (  # noqa: F401
+    _assert_index_available,
+    _index_path,
+    _read_index,
+    _remove_index,
+    _write_index,
+)
 
 def _validate_anchor_identity(
     lineage: ManagedHeadlessSessionLineage,
